// src/lib/api/generated/model/recordingStatusSchema.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The canonical backend API for COMPASS clients.
 * OpenAPI spec version: 1.0.0
 */
import type { RecordingRunProjectionSchema } from './recordingRunProjectionSchema';

export interface RecordingStatusSchema {
  reference_code: string;
  run?: RecordingRunProjectionSchema | null;
  status: string;
}